Factor out shared code from start and stop
