+++ cmd/soroban-cli/src/commands/network/stop.rs
@@ -1,41 +1,10 @@
-use core::fmt;
-
-use bollard::{ClientVersion, Docker};
-use clap::ValueEnum;
+use crate::commands::network::shared::connect_to_docker;
+use crate::commands::network::shared::Network;
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
     #[error("Failed to stop container: {0}")]
     StopContainerError(#[from] bollard::errors::Error),
-}
-
-// DEFAULT_TIMEOUT and API_DEFAULT_VERSION are from the bollard crate
-const DEFAULT_TIMEOUT: u64 = 120;
-const API_DEFAULT_VERSION: &ClientVersion = &ClientVersion {
-    major_version: 1,
-    minor_version: 40,
-};
-
-// TODO: move to a shared module
-#[derive(ValueEnum, Debug, Clone, PartialEq)]
-pub enum Network {
-    Local,
-    Testnet,
-    Futurenet,
-    Pubnet,
-}
-
-impl fmt::Display for Network {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let variant_str = match self {
-            Network::Local => "local",
-            Network::Testnet => "testnet",
-            Network::Futurenet => "futurenet",
-            Network::Pubnet => "pubnet",
-        };
-
-        write!(f, "{}", variant_str)
-    }
 }
 
 #[derive(Debug, clap::Parser, Clone)]
@@ -51,20 +20,10 @@ pub struct Cmd {
 impl Cmd {
     pub async fn run(&self) -> Result<(), Error> {
         let container_name = format!("stellar-{}", self.network);
-        let docker = connect_to_docker(self);
+        let docker = connect_to_docker(&self.docker_socket_path);
         println!("Stopping container: {container_name}");
         docker.stop_container(&container_name, None).await.unwrap();
 
         Ok(())
-    }
-}
-
-//TODO: move to a shared module
-fn connect_to_docker(cmd: &Cmd) -> Docker {
-    if cmd.docker_socket_path.is_some() {
-        let socket = cmd.docker_socket_path.as_ref().unwrap();
-        Docker::connect_with_socket(socket, DEFAULT_TIMEOUT, API_DEFAULT_VERSION).unwrap()
-    } else {
-        Docker::connect_with_socket_defaults().unwrap()
     }
 }
